fix: MyPy batch 10 (88 errors from 10 files)
Fixed MyPy type checking errors in batch 10 across 10 files:

- victor/workflows/team_collaboration.py:
  - Added type annotations for send_and_log function
  - Fixed tuple unpacking with type checking for exception handling
  - Changed total_response_time to float type

- victor/workflows/team_dashboard_server.py:
  - Fixed type: ignore comment to specify arg-type
  - Added return type annotations for health_check and websocket_endpoint
  - Added None checks for datetime.isoformat() calls

- victor/workflows/team_node_runner.py:
  - Removed invalid timeout_seconds parameter from execute_task call
  - Fixed SubAgentRole.ASSISTANT to SubAgentRole.EXECUTOR
  - Fixed ObservabilityBus.get_instance() to instantiate and use emit()

- victor/workflows/team_templates.py:
  - Added type annotations for **overrides parameters (Any)

- victor/workflows/team_tracing.py:
  - Fixed class-level lock type from threading.Lock to threading.RLock

- victor/workflows/unified_compiler.py:
  - Added orchestrator None check for team node execution
  - Added type parameters to generic types (CompiledGraph, GraphExecutionResult, etc.)
  - Fixed SubAgentOrchestrator instantiation with AgentOrchestrator cast
  - Changed execute_task to spawn method with proper return type handling
  - Added getattr for continue_on_error attribute with default False
  - Added cast() for retry result returns
  - Added tuple type annotations for parallel execution
  - Added exception type checking for parallel results
  - Fixed get_tool_registry() to use ServiceContainer
  - Added workflow_name default value handling

- victor/workflows/unified_executor.py:
  - Fixed get_tool_registry() to use ServiceContainer

- victor/workflows/validator.py:
  - Imported LinterResult type
  - Changed ValidationResult to LinterResult in type annotations
  - Added None check for result

- victor/workflows/visualization.py:
  - Fixed HITLNode type with Optional[Any] and proper import
  - Added cast() for method calls in SVG rendering
  - Added Callable cast for PNG rendering methods
  - Added str() cast for svg_content return

- victor/workflows/yaml_loader.py:
  - Added type annotations for lambda function parameters
  - Added List[tuple[str, str]] type for auto_workflows
  - Added Callable type annotation for type_map
  - Added type: ignore for converter call
  - Fixed VolumeMount parameters (source, target instead of host_path, container_path)
  - Fixed HealthCheckConfig path parameter with default
  - Fixed LifecycleConfig (shutdown_grace instead of shutdown_timeout)
  - Added type: ignore for ServiceConfig provider and command
  - Added Dict[str, Any] type annotations for nodes
  - Made condition_registry and transform_registry Optional

+++ victor/optimizations/memory.py
@@ -151,15 +151,15 @@ class ObjectPool(Generic[T]):
         with self._lock:
             self._pool.clear()
 
-    def get_stats(self) -> Dict[str, int]:
+    def get_stats(self) -> Dict[str, float]:
         """Get pool statistics."""
         with self._lock:
             return {
-                "pool_size": len(self._pool),
-                "max_size": self._max_size,
-                "created": self._created,
-                "acquired": self._acquired,
-                "reused": self._reused,
+                "pool_size": float(len(self._pool)),
+                "max_size": float(self._max_size),
+                "created": float(self._created),
+                "acquired": float(self._acquired),
+                "reused": float(self._reused),
                 "reuse_rate": float(self._reused / self._acquired if self._acquired > 0 else 0.0),
             }
 
@@ -405,7 +405,7 @@ class MemoryOptimizer:
         stats.gc_counts = gc.get_count()
 
         # Get pool stats
-        stats.pool_stats = {name: pool.get_stats() for name, pool in self._pools.items()}
+        stats.pool_stats = {name: pool.get_stats() for name, pool in self._pools.items()}  # type: ignore[assignment]
 
         return stats
